通用響應 instagram4J登入用controller

+++ src/main/java/org/example/dao/LoginAccountDao.java
@@ -3,7 +3,5 @@ package org.example.dao;
 import org.example.entity.LoginAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface LoginAccountDao extends JpaRepository<LoginAccount, BigInteger> {
+public interface LoginAccountDao extends JpaRepository<LoginAccount, Integer> {
 }
