+++ src/main/java/org/example/service/BaseService.java
@@ -1,6 +1,5 @@
 package org.example.service;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,7 +7,7 @@ public interface BaseService<T> {
 
     Optional<T> save(T target);
 
-    Optional<T> findById(BigInteger id);
+    Optional<T> findById(Integer id);
 
     List<T> findAll();
 }
